fixed after eval

+++ philo/sim_mutex.c
@@ -6,7 +6,7 @@
 /*   By: ccraciun <ccraciun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 15:18:46 by ccraciun          #+#    #+#             */
-/*   Updated: 2024/10/23 15:22:01 by ccraciun         ###   ########.fr       */
+/*   Updated: 2024/11/01 11:53:29 by ccraciun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,27 +19,11 @@
 */
 static int	init_philosopher_mutexes(t_philosopher *philosopher)
 {
-	philosopher->time_zero_mut = ft_calloc(1, sizeof(pthread_mutex_t));
 	philosopher->last_meal_mut = ft_calloc(1, sizeof(pthread_mutex_t));
-	philosopher->sim_stop_mut = ft_calloc(1, sizeof(pthread_mutex_t));
-	if (!philosopher->time_zero_mut || !philosopher->last_meal_mut
-		|| !philosopher->sim_stop_mut)
-		return (1);
-	if (pthread_mutex_init(philosopher->time_zero_mut, NULL) != 0)
-	{
-		ft_error("Failed to init time_zero mutex\n");
-		return (1);
-	}
+	if (!philosopher->last_meal_mut)
+		return (ft_error("Calloc failed\n"));
 	if (pthread_mutex_init(philosopher->last_meal_mut, NULL) != 0)
-	{
-		ft_error("Failed to init last_meal mutex\n");
-		return (1);
-	}
-	if (pthread_mutex_init(philosopher->sim_stop_mut, NULL) != 0)
-	{
-		ft_error("Failed to init sim_stop mutex\n");
-		return (1);
-	}
+		return (ft_error("Failed to init last_meal mutex\n"));
 	return (0);
 }
 
@@ -51,28 +35,24 @@ static int	init_philosopher_mutexes(t_philosopher *philosopher)
 */
 static int	destroy_init_mutexes(t_simulation *sim, int index)
 {
-	int	status;
 	int	i;
 
-	status = 0;
-	if (pthread_mutex_destroy(&sim->print_mutex) != 0)
-	{
-		ft_error("Failed to destroy print mutex\n");
-		status = 1;
-	}
 	i = index;
+	if (pthread_mutex_destroy(&sim->print_mutex) != 0)
+		return(ft_error("Failed to destroy print mutex\n"));
+	if (pthread_mutex_destroy(&sim->sim_stop_mut) != 0)
+		return(ft_error("Failed to destroy sim_stop_mut mutex\n"));
+	if (pthread_mutex_destroy(&sim->time_zero_mut) != 0)
+		return(ft_error("Failed to destroy time_zero_mut mutex\n"));
 	while (i >= 0)
 	{
 		if (pthread_mutex_destroy(&sim->forks[i]) != 0)
-		{
-			ft_error("Failed to destroy fork mutex\n");
-			status = 1;
-		}
+			return(ft_error("Failed to destroy fork mutex\n"));
 		if (destroy_philosopher_mutexes(&sim->philosophers[i]) != 0)
-			status = 1;
+			return(ft_error("Failed to destroy philo mutexes\n"));
 		i--;
 	}
-	return (status);
+	return (0);
 }
 
 /*
@@ -84,24 +64,24 @@ int	init_mutexes(t_simulation *sim)
 {
 	int	i;
 
+	i = 0;
 	if (pthread_mutex_init(&sim->print_mutex, NULL) != 0)
 		return (free_sim_memory(sim), 1);
-	i = 0;
+	if (pthread_mutex_init(&sim->sim_stop_mut, NULL) != 0)
+		return (pthread_mutex_destroy(&sim->print_mutex),
+			free_sim_memory(sim), 1);
+	if (pthread_mutex_init(&sim->time_zero_mut, NULL) != 0)
+		return (pthread_mutex_destroy(&sim->print_mutex),
+			pthread_mutex_destroy(&sim->sim_stop_mut),
+			free_sim_memory(sim), 1);
 	while (i < sim->num_philosophers)
 	{
 		if (init_philosopher_mutexes(&sim->philosophers[i]) != 0)
-		{
-			destroy_init_mutexes(sim, i - 1);
-			free_sim_memory(sim);
-			return (2);
-		}
+			return (destroy_init_mutexes(sim, i - 1), free_sim_memory(sim), 2);
 		if (pthread_mutex_init(&sim->forks[i], NULL) != 0)
-		{
-			ft_error("Failed to init fork mutex\n");
-			destroy_init_mutexes(sim, i);
-			free_sim_memory(sim);
-			return (2);
-		}
+			return (ft_error("Failed to init fork mutex\n"),
+					destroy_init_mutexes(sim, i),
+					free_sim_memory(sim), 2);
 		i++;
 	}
 	return (0);
